pyproject.toml
[project]
name = "intro-to-rag"
version = "0.1.0"
description = ""
authors = [
    {name = "pdemeulenaer",email = "pdemeulenaer@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "unstructured[all-docs] (>=0.16.23,<0.17.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "lxml (>=5.3.1,<6.0.0)",
    "chromadb (>=0.6.3,<0.7.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "langchain (>=0.3.20,<0.4.0)",
    "langchain-community (>=0.3.19,<0.4.0)",
    "langchain-openai (>=0.3.7,<0.4.0)",
    "langchain-groq (>=0.2.4,<0.3.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "streamlit (>=1.43.2,<2.0.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "altair (>=5.5.0,<6.0.0)",
    "sentence-transformers (>=3.4.1,<4.0.0)"
]

[tool.poetry]
packages = [{include = "intro_to_rag", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
